Refactor Kubernetes keystore into its own file.
This will allow GCP/AWS Secret Manager implementations to sensibly live
in their own files, too.
